refactor code for readibility
